Add files via upload
1) The ZProbe function has been created for lowering and raising the probe.
2) In the calibration commands G29, G30, G31, G32 added automatic lowering and raising of the probe.
3) An optional "nozzle_rise" setting has been added to the "zprobe" section of the "config.txt" file, allowing you to specify the offset of the probe of the table level sensor relative to the nozzle level along the "Z" axis. The value is indicated in millimeters (mm). If the sensor working height is below the nozzle level, then the value should be indicated as positive.
4) Added additional functionality to the G30 table search command:
a) "G30 Z0" sets Z=0 at the table level (that is, at the sensor triggering level, taking into account the value specified in the "nozzle_rise" parameter);
b) "G30" with parameter "D" ignores the modifications made.

Lowering and raising the probe in commands G29, G30, G31, G32 is performed by sending commands "M280" and "M281".

Example: The table level sensor is triggered when a distance of 0.67 mm remains between the table and the nozzle (that is, the sensor response height is 0.67 mm below the nozzle level). The file "config.txt" should set
zprobe.nozzle_rise 0.67
In this case:
i) The “G30 Z0” command sets the coordinate value Z = 0 + 0.67 = 0.67 when the sensor is triggered. After that, the “G1 Z0” command will set the table to the nozzle level.
ii) The “G30 Z0 D” command sets Z=0 at the sensor trigger height;
iii) The “G30 D” command searches for the table before the sensor is triggered, displays the Z value at the moment of operation and rolls the table back to its original height.